trading_ai/requirements.txt
# Core
pydantic==2.7.1
pydantic-settings==2.3.4

# Data
ccxt==4.3.20
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b

# ML
xgboost==2.0.3
scikit-learn==1.5.0

# LLM
anthropic==0.28.0

# API
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Telegram
python-telegram-bot==21.3
